Moved readSettingsFile method from ServerMain.java to ServerSettings.java
