+++ .gitignore
@@ -12,3 +12,4 @@ server*.jar
 *~
 errorLog.txt
 *.properties
+out
