+++ src/server/ServerSettings.java
@@ -1,9 +1,9 @@
 package server;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import database.constants.ServerDependentValues;
+
+import java.io.*;
+import java.util.Scanner;
 
 public class ServerSettings {
 
@@ -69,4 +69,61 @@ public class ServerSettings {
 		}
 	}
 
+	public static void readSettingsFile(String path) throws FileNotFoundException {
+		File dbFile = new File(path);
+		if (dbFile.exists()) {
+			Scanner scan = new Scanner(dbFile);
+			while (scan.hasNextLine()) {
+				String line = scan.nextLine();
+				int index = line.indexOf("=");
+				String key = line.substring(0, index).trim();
+				String value = line.substring(index+1).trim();
+				switch (key.toLowerCase()) {
+				case "databaseuser":
+					databaseUsername = value;
+					break;
+				case "databasepassword":
+					databasePassword = value;
+					break;
+				case "databasehost":
+					databaseHost = value;
+					break;
+				case "databasename":
+					databaseName = value;
+					break;
+				case "publicaddress":
+					publicAddress = value;
+					break;
+				case "apacheport":
+					apachePort = Integer.parseInt(value);
+					break;
+				case "downloadurl":
+					downloadURL = value;
+					break;
+				case "uploadurl":
+					uploadURL = value;
+					break;
+				case "genomizerport":
+					genomizerPort = Integer.parseInt(value);
+					break;
+				case "passwordhash":
+					passwordHash = value;
+					break;
+				case "passwordsalt":
+					passwordSalt = value;
+					break;
+				default:
+					System.err.println("Unrecognized setting: " + key);
+					break;
+				}
+			}
+			scan.close();
+			ServerDependentValues.DownloadURL = publicAddress + ":" +
+					apachePort + downloadURL;
+			ServerDependentValues.UploadURL = publicAddress + ":" +
+					apachePort + uploadURL;
+		} else {
+			System.err.println("Error, " + path + " does not exist, using default settings.");
+		}
+	}
 }
